HACK: Improve diffs between patch sets in a change.
By filtering out changes that are purely due to differences between
the base commits. This has a big effect when patch sets were rebased.

This is a hack because it makes some constructors public that used to
be private and doesn't make the filtering optional somehow.
It should be safe, however, as it prefers to not drop an edit when in
doubt.